requirements.txt
psycopg2
questionary
rich
python-dotenv
mysql-connector-python
